+++ lib/utils/value_notifier.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 
 extension ValueNotifierExtension<T> on ValueNotifier<T?> {
   Future<T> waitForValue() async {
+    if (value != null) return value!;
     final completer = Completer<T>();
     void listener() {
       if (value != null) {