+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT 
 
 - Fixes the refresh and stays on the same URL.
+- Only allow the access to the user screen to the admin.
 
 ## 0.14.0 [2023-10-22]